+++ lib/src/feedback_manager/dialog_feedback_mixin.dart
@@ -2,6 +2,7 @@ import 'package:cubes/src/feedback_manager/feedback_manager.dart';
 import 'package:cubes/src/observable/observable_value.dart';
 import 'package:flutter/material.dart';
 
+/// Class responsible for configuring the dialogs
 class CDialogController<T> {
   final bool dismissible;
   final Color barrierColor;
@@ -26,6 +27,7 @@ class CDialogController<T> {
   });
 }
 
+/// Mixin responsible for adding listeners to ObservableValue and controlling the display of dialogs
 mixin DialogFeedBackMixin<T extends StatefulWidget> on State<T> {
   static const ANIMATION_DURATION = 150;
   Map<CDialogController, bool> _mapDialogIsShowing = Map();
@@ -61,17 +63,18 @@ mixin DialogFeedBackMixin<T extends StatefulWidget> on State<T> {
     await Future.delayed(Duration(milliseconds: ANIMATION_DURATION));
     _mapDialogIsShowing[element] = true;
     await showDialog(
-      context: context,
-      barrierDismissible: element.dismissible,
-      barrierColor: element.barrierColor,
-      useSafeArea: element.useSafeArea,
-      useRootNavigator: element.useRootNavigator,
-      routeSettings: element.routeSettings,
-      child: WillPopScope(
-        onWillPop: () => Future.value(element.dismissible),
-        child: element.doBuild(element.observable.value.data, context),
-      ),
-    );
+        context: context,
+        barrierDismissible: element.dismissible,
+        barrierColor: element.barrierColor,
+        useSafeArea: element.useSafeArea,
+        useRootNavigator: element.useRootNavigator,
+        routeSettings: element.routeSettings,
+        builder: (context) {
+          return WillPopScope(
+            onWillPop: () => Future.value(element.dismissible),
+            child: element.doBuild(element.observable.value.data, context),
+          );
+        });
     _mapDialogIsShowing[element] = false;
     // ignore: invalid_use_of_protected_member
     element.observable.setInitialValue(element.observable.value.copyWith(show: false));