+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.10.3
+
+- [BREAKING CHANGE] renamed `ready()` to `onReady(Object arguments)` in `Cube`;
+- renamed `initData` to `arguments` in `CubeWidget` and `CubeBuilder`;
+- update get_it to `5.0.4`;
+
 # 0.10.2
 
 - improvements in  `CAnimatedList`.